# archive_old/docs/swagger.yaml
basePath: /api/v1/
definitions:
  model.Campaign:
    properties:
      category:
        type: string
      collected_amount:
        type: number
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      id:
        type: integer
      min_donation:
        type: number
      status:
        type: string
      target_amount:
        type: number
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.Donation:
    properties:
      amount:
        type: number
      campaign:
        $ref: '#/definitions/model.Campaign'
      campaign_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a crowdfunding API.
  title: Crowdfunding API
  version: "1.0"
paths:
  /donations:
    post:
      consumes:
      - application/json
      description: Create a new donation for a user
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Donation object
        in: body
        name: model.Donation
        required: true
        schema:
          $ref: '#/definitions/model.Donation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create a new donation
      tags:
      - donations
  /donations/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific donation by its ID
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Donation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get Donation details by Donation ID
      tags:
      - donations
    put:
      consumes:
      - application/json
      description: Get details of a specific donation by its ID
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Donation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Donation object
        in: body
        name: model.Donation
        required: true
        schema:
          $ref: '#/definitions/model.Donation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update a donation based on the invoice status
      tags:
      - donations
...
swagger: "2.0"